wee-2011-06-27 import. - 2011-06-27 (tinybit)In grub4dos.h, added API wee_skip_to, added the 2nd argument of "int" for API print_a_completion. - 2011-06-26 (tinybit)corrected the return data type of grub_read from 32bit to 64bit to match the defination in grub4dos.h. also updated grub4dos.h.
git-svn-id: http://grubutils.googlecode.com/svn@82 d95a8657-0e37-07c3-0af6-720afd5d0260

+++ grubutils/wee/shared.h
@@ -553,7 +553,7 @@ int grub_open (char *filename);
 
 /* Read LEN bytes into BUF from the file that was opened with
    GRUB_OPEN.  If LEN is -1, read all the remaining data in the file.  */
-unsigned long grub_read (unsigned long long buf, unsigned long long len, unsigned long write);
+unsigned long long grub_read (unsigned long long buf, unsigned long long len, unsigned long write);
 
 /* Display statistics on the current active device. */
 void print_fsys_type (void);
